problem in sorting large arrays, fixed transform function, buffer size is increased

+++ src/server/OutputManager.java
@@ -12,8 +12,8 @@ public class OutputManager {
     public static String outputManager(String line){
         int[] unsortedArray = MessageTransform.transform(line);
         StringBuilder buildAnswer = new StringBuilder("Sorted array: ");
-        long timeSequential = TimerSequentially.TimerSeqSorting(unsortedArray);
-        long timeParallel = TimerParallel.TimerParallelSorting(unsortedArray);
+        double timeSequential = TimerSequentially.TimerSeqSorting(unsortedArray);
+        double timeParallel = TimerParallel.TimerParallelSorting(unsortedArray);
         buildAnswer.append(Arrays.toString(unsortedArray)).append('\n');
         buildAnswer.append("Sequentially sorted time: ").append(timeSequential).append('\n');
         buildAnswer.append("Parallel sorted time: ").append(timeParallel).append('\n');
